+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Bumped version
 - endpoint to retrieve product by id in `src/api/v1/apss/orders/routes.py`
 - template for update/delete product in `src/templates`
 - test for enppoint orders/products/{id} in `tests/integration/test_products.py`
+- test for template products/updte in `tests/integration/test_update_products.py`
 
 ### Changed
 - N/A